canny edge detection

+++ Engine/core/graphic/renderPass/cannyEdgeDetection/sobel/sobelPass.cpp
@@ -18,7 +18,7 @@ SobelPass::SobelPass(Renderer* r, const RenderState& state)
 	{
 		{
 			AttachmentType::Color,
-			TextureInternalFormat::RGB8,
+			TextureInternalFormat::RGB16F,
 			TextureDataFormat::RGB,
 			TextureWarpMode::CLAMP_TO_EDGE,
 			TextureWarpMode::CLAMP_TO_EDGE,
